公寓字典模型建立

+++ SmartApartment/SetNewPwdController.m
@@ -141,7 +141,7 @@
 - (NSString *)md5:(NSString *)str{
     const char *cStr = [str UTF8String];
     unsigned char result[16];
-    CC_MD5(cStr, strlen(cStr), result); // This is the md5 call
+    CC_MD5(cStr, (unsigned int)strlen(cStr), result); // This is the md5 call
     return [NSString stringWithFormat:
             @"%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x",
             result[0], result[1], result[2], result[3],
